+++ src/tewi/service/search/__init__.py
@@ -3,6 +3,7 @@ from .yts_provider import YTSProvider
 from .torrentscsv_provider import TorrentsCsvProvider
 from .tpb_provider import TPBProvider
 from .nyaa_provider import NyaaProvider
+from .jackett_provider import JackettProvider
 
 __all__ = ['BaseSearchProvider', 'YTSProvider', 'TorrentsCsvProvider',
-           'TPBProvider', 'NyaaProvider']
+           'TPBProvider', 'NyaaProvider', 'JackettProvider']