+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for Deluge torrent client (via Web API) [#110](https://github.com/anlar/tewi/issues/110)
 - Configuration profiles support (`--profile` option to load profile configs, `--profiles` to list available profiles) [#108](https://github.com/anlar/tewi/issues/108)
+- Add Jackett torrent search provider [#118](https://github.com/anlar/tewi/issues/118)
+- Configuration options `jackett_url` and `jackett_api_key` in [search] section
+- Support for HTTP/HTTPS torrent file URLs in search results
+- New `torrent_link` field in `SearchResultDTO` for torrent file URLs
+- Automatic download and conversion of torrent files from URLs in Deluge client
 
 ### Changed